feature:发布帖子

+++ app/src/main/java/com/blinker/video/model/TagList.kt
@@ -18,6 +18,6 @@ data class TagList(
     val icon: String?,
     val id: Int,
     val intro: String?,
-    val tagId: Int,
+    val tagId: Long,
     val title: String?
 )
